EIF: set the flag that says we're NX-compatible
PE defines a flag in the DLL characteristics to say that a binary is
compatible with being loaded with NX (W^X) mitigations enabled.  When
the mitigation is enabled, the stack is not executable, and
AllocatePages() and AllocatePool() return memory that is set +rw -x.

As far as I can tell, memtest86+ is completely compatible with this, so
this patch sets that flag.

Signed-off-by: Peter Jones <pjones@redhat.com>