Busy loop calibration
Non-TSC capable systems rely on the busy loop that currently assumed 1000 loops
per 1 us.  In case of older system such as 486, this is highly overestimated.

Add code for busy loop calibration to provide a better number than 1000.
As we are not allowed to use interrupts, the precision is not perfect,
but should provide a number that is close enough, for example 7 on my
486 DX4-120.