add rtc_result commandline option for automation
When set and an automatic reboot due to the "autoreboot" option should be triggered,
the date in the hardware real time clock (RTC) is set to 2012-01-01. This acts
as indicator that the memtest really was successful and the reboot did not occur
due to faulty hardware, power loss or similar things.

Includes extensive documentation about automating Memtest86+ in README_automation.md