WIP BROKEN Work on NUMA awareness (#12): parse the ACPI SRAT to build up new internal structures related to proximity domains and affinity; use these structures in setup_vm_map() and calculate_chunk() to skip the work on the processors which don't belong to the proximity domain currently being tested.
Tested on QEMU at first; a short test of a previous iteration of the algorithm on a real 2S Westmere server showed the consequence of two related mistakes in the second top-level loop of find_numa_nodes_in_srat(), though no deadlock. Now tested on both the 2S Westmere and two 4S Opteron server models.

The parsing code works, the new data structures are built as they should be.
FIXME: a deadlock issue was "solved" by calling more do_tick(), but that skews the progress indicator a lot.
FIXME: the code is broken for a single NUMA node, it spits huge amounts of errors... the NUMA emulation code must be buggy.