Add "singlepass" option
Normally memtest86+ runs forever. Using this command line option
causes it to exit/reboot after one pass of the selected tests is
performed.

Also adds the feature as a menu option and if enabled from the menu
during testing it will cause the memtest86+ to exit/reboot after
the current pass is completed.