Add a tool allowing to (re)generate the jep106.S file
generate_jep106.py can be used to parse JEP106 PDF files from JEDEC and
to directly generate / update the `jep106.S` file.

If an existing jep106.S file is specified, the new entries are added
as disabled, the enable / disable state for existing entries is
maintained, and names for the currently enabled ones are preserved.

The next time we have to do the update, it can be done with minimal
effort. The parser was designed with robustness in mind and has been
tested successfully on the following JEDEC documents:
 - JEP106-K from 2001-09-13
 - JEP106AB from 2010-02-23
 - JEP106AJ from 2012-08-17
 - JEP106BH from 2023-05-01
 - JEP106BJ from 2024-01-04
 - JEP106BK from 2024-09-17