Don't pad test descriptions
There is no need to pad test descriptions with spaces and embed them
with brackets, use printf_region instead.

With this change, we can also save 350 bytes by changing the description
field in test_pattern_t to be declared as a pointer, which will also
move it from .data to .rodata.