Add USB hub quirk to support AMI virtual keyboard (issue #523) (#535)
* Add USB hub quirk to support AMI virtual keyboard (issue #523)

From testing it appears that the AMI virtual devices do not work
correctly if you disable one port on the virtual hub then reuse the
USB address for a device connected to another port on that hub. So
add a quirk to detect the virtual hub from its vendor/product ID
and leave all ports on that hub active when performing the scan for
USB keyboards. This will permanently reserve a bit more memory, but
not significantly so.

* Replace hard-coded USB vendor ID with macro definition.