Serial/TTY: redraw in initial and reboot cases
During test execution, do_tick() triggers redrawing the TTY 'screen'. But
there are cases before and after that where the TTY is not updated.

For example, the initial interface nor the 'initial notice' giving the user
the option to configure before starting the test, is not displayed. The
screen remains blank until the tests start.
Similarly, the 'Rebooting...' notices are not shown.

Add a number of 'zoned' tty_*_redraw() functions and call them in the right
places to fix this.